requirements.txt
google-generativeai==0.3.2
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
pyyaml==6.0.1
pytest==7.4.0
python-dotenv==1.0.0
tabulate==0.9.0
python-dateutil==2.8.2
